Break `Account` and `MuxedAccount` apart into separate files
